+++ R/consecutive_mutate_linter.R
@@ -73,11 +73,8 @@ consecutive_mutate_linter <- function(invalid_backends = "dbplyr") {
 
   Linter(function(source_expression) {
     # need the full file to also catch usages at the top level
-    if (!is_lint_level(source_expression, "file")) {
-      return(list())
-    }
-
     xml <- source_expression$full_xml_parsed_content
+    if (is.null(xml)) return(list())
 
     attach_str <- get_r_string(xml_find_all(xml, attach_pkg_xpath))
     if (any(invalid_backends %in% attach_str)) {
@@ -97,5 +94,5 @@ consecutive_mutate_linter <- function(invalid_backends = "dbplyr") {
       lint_message = "Unify consecutive calls to mutate().",
       type = "warning"
     )
-  })
+  }, linter_level = "file")
 }