+++ R/expect_named_linter.R
@@ -42,16 +42,13 @@ expect_named_linter <- function() {
   "
 
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "expression")) {
-      return(list())
-    }
-
     xml <- source_expression$xml_parsed_content
+    if (is.null(xml)) return(list())
 
     bad_expr <- xml_find_all(xml, xpath)
     matched_function <- xp_call_name(bad_expr)
     lint_message <- sprintf("expect_named(x, n) is better than %s(names(x), n)", matched_function)
 
     xml_nodes_to_lints(bad_expr, source_expression = source_expression, lint_message, type = "warning")
-  })
+  }, linter_level = "expression")
 }