+++ R/expect_type_linter.R
@@ -60,11 +60,8 @@ expect_type_linter <- function() {
   xpath <- paste(expect_equal_identical_xpath, "|", expect_true_xpath)
 
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "expression")) {
-      return(list())
-    }
-
     xml <- source_expression$xml_parsed_content
+    if (is.null(xml)) return(list())
 
     bad_expr <- xml_find_all(xml, xpath)
     matched_function <- xp_call_name(bad_expr)
@@ -79,5 +76,5 @@ expect_type_linter <- function() {
       lint_message = msg,
       type = "warning"
     )
-  })
+  }, linter_level = "expression")
 }