+++ R/extraction_operator_linter.R
@@ -62,11 +62,8 @@ extraction_operator_linter <- function() {
   ")
 
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "expression")) {
-      return(list())
-    }
-
     xml <- source_expression$xml_parsed_content
+    if (is.null(xml)) return(list())
     bad_exprs <- xml_find_all(xml, xpath)
     msgs <- sprintf("Use `[[` instead of `%s` to extract an element.", xml_text(bad_exprs))
 
@@ -76,5 +73,5 @@ extraction_operator_linter <- function() {
       lint_message = msgs,
       type = "warning"
     )
-  })
+  }, linter_level = "expression")
 }