+++ R/fixed_regex_linter.R
@@ -140,11 +140,8 @@ fixed_regex_linter <- function(allow_unescaped = FALSE) {
   ")
 
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "expression")) {
-      return(list())
-    }
-
     xml <- source_expression$xml_parsed_content
+    if (is.null(xml)) return(list())
 
     patterns <- xml_find_all(xml, xpath)
     pattern_strings <- get_r_string(patterns)
@@ -171,5 +168,5 @@ fixed_regex_linter <- function(allow_unescaped = FALSE) {
       lint_message = msg,
       type = "warning"
     )
-  })
+  }, linter_level = "expression")
 }