+++ R/namespace_linter.R
@@ -40,11 +40,8 @@
 #' @export
 namespace_linter <- function(check_exports = TRUE, check_nonexports = TRUE) {
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "file")) {
-      return(list())
-    }
-
     xml <- source_expression$full_xml_parsed_content
+    if (is.null(xml)) return(list())
 
     ns_nodes <- xml_find_all(xml, "//NS_GET | //NS_GET_INT")
 
@@ -120,7 +117,7 @@ namespace_linter <- function(check_exports = TRUE, check_nonexports = TRUE) {
     }
 
     lints
-  })
+  }, linter_level = "file")
 }
 
 namespace_symbols <- function(ns, exported = TRUE) {