+++ R/nested_ifelse_linter.R
@@ -87,11 +87,8 @@ nested_ifelse_linter <- function() {
   ")
 
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "expression")) {
-      return(list())
-    }
-
     xml <- source_expression$xml_parsed_content
+    if (is.null(xml)) return(list())
 
     bad_expr <- xml_find_all(xml, xpath)
 
@@ -101,5 +98,5 @@ nested_ifelse_linter <- function() {
       "instead, try (1) data.table::fcase; (2) dplyr::case_when; or (3) using a lookup table."
     )
     xml_nodes_to_lints(bad_expr, source_expression, lint_message, type = "warning")
-  })
+  }, linter_level = "expression")
 }