+++ R/nested_pipe_linter.R
@@ -68,11 +68,8 @@ nested_pipe_linter <- function(
   ")
 
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "expression")) {
-      return(list())
-    }
-
     xml <- source_expression$xml_parsed_content
+    if (is.null(xml)) return(list())
 
     bad_expr <- xml_find_all(xml, xpath)
 
@@ -82,5 +79,5 @@ nested_pipe_linter <- function(
       lint_message = "Don't nest pipes inside other calls.",
       type = "warning"
     )
-  })
+  }, linter_level = "expression")
 }