+++ R/one_call_pipe_linter.R
@@ -66,11 +66,8 @@ one_call_pipe_linter <- function() {
   ")
 
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "expression")) {
-      return(list())
-    }
-
     xml <- source_expression$xml_parsed_content
+    if (is.null(xml)) return(list())
 
     bad_expr <- xml_find_all(xml, xpath)
     pipe <- xml_find_chr(bad_expr, "string(SPECIAL | PIPE)")
@@ -81,5 +78,5 @@ one_call_pipe_linter <- function() {
       lint_message = paste0("Expressions with only a single call shouldn't use pipe ", pipe, "."),
       type = "warning"
     )
-  })
+  }, linter_level = "expression")
 }