+++ R/pipe_call_linter.R
@@ -27,11 +27,8 @@ pipe_call_linter <- function() {
   xpath <- glue("//SPECIAL[{ xp_text_in_table(pipes) }]/following-sibling::expr[*[1][self::SYMBOL]]")
 
   Linter(function(source_expression) {
-    if (!is_lint_level(source_expression, "expression")) {
-      return(list())
-    }
-
     xml <- source_expression$xml_parsed_content
+    if (is.null(xml)) return(list())
 
     bad_expr <- xml_find_all(xml, xpath)
     pipe <- xml_text(xml_find_first(bad_expr, "preceding-sibling::SPECIAL[1]"))
@@ -43,5 +40,5 @@ pipe_call_linter <- function() {
         sprintf("Use explicit calls in magrittr pipes, i.e., `a %1$s foo` should be `a %1$s foo()`.", pipe),
       type = "warning"
     )
-  })
+  }, linter_level = "expression")
 }